Added what could happen if you give false information.
I checked with the divorce team on what they're using. It's very likely we'll have to use the same statement of truth on all services. I added "You could be fined or imprisoned if you deliberately submit false information". The divorce service includes " for contempt of court" - they haven't been able to find a plain English version. I don't think we need that term, but the lawyers will tell us.